requirements.txt
Flask==2.3.0
Flask-Login==0.6.2
pymongo==4.3.3
python-dotenv==1.0.0
Werkzeug==2.3.0
gunicorn==20.1.0
Pillow==9.5.0
python-multipart==0.0.6
